scrape/requirements.txt
python-dotenv
requests
beautifulsoup4
google-cloud-storage
eventregistry